# Conflicts:
#	app/src/components/layout/ImageCards.vue
#	app/src/views/Home.vue

+++ etl/scripts/run_etl.sh
@@ -1,18 +1,42 @@
 #!/bin/bash
-#changing working dir
-cd /home/etl/einander-helfen/etl/scripts
+lockfile=./ETLlock
 
-# Installing required python packages
-pip3 install -r ../requirements.txt
+if ( set -o noclobber; echo "$$" > "$lockfile") 2> /dev/null; then
 
-# executing crawl and enhancement
-python3 ../main.py
+        trap 'rm -f "$lockfile"; exit $?' INT TERM EXIT
 
-# package enhanced data for prod
-tar cfvz enhanced_output.tar.gz ../data_enhancement/data ../data_enhancement/output
+        # Default context
+        context=DE
 
-# copy packaged data to make it available for nginx endpoint
-cp enhanced_output.tar.gz /var/www/html
+        while getopts :hc: flag
+        do
+          case "${flag}" in
+            h) echo "Switch context with flag -c ['DE' (Default), 'US']";exit;;
+            c) context=${OPTARG};;
+          esac
+        done
 
-# execute elastic upload
-python3 ../execute_elastic_upload.py
+        #changing working dir
+        cd /home/etl/einander-helfen/etl/scripts
+
+        # Installing required python packages
+        pip3 install -r ../requirements.txt
+
+        # executing crawl and enhancement
+        python3 ../main.py -c "$context"
+
+        # package enhanced data for prod
+        tar cfvz enhanced_output.tar.gz ../data_enhancement/data ../data_enhancement/output
+
+        # copy packaged data to make it available for nginx endpoint
+        cp enhanced_output.tar.gz /var/www/html
+
+        # execute elastic upload
+        python3 ../execute_elastic_upload.py -c "$context"
+
+        # clean up after yourself, and release your trap
+        rm -f "$lockfile"
+        trap - INT TERM EXIT
+else
+        echo "Lock Exists: $lockfile owned by $(cat $lockfile)"
+fi
